requirements.txt
Flask==2.0.1
numpy>=1.24.0
Werkzeug==2.0.1
click==8.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
flask-cors==4.0.0
flask_cors
requests==2.31.0
beautifulsoup4==4.12.0
python-dotenv==1.0.0
google-generativeai>=0.3.0
